Handle issues with consul leader node failures
If the consul leader node goes down, node down notification is not added
in the consul key value store. When a new node is elected as a leader
it used to populate the cluster status from key value store. With this
change when a new leader is elected it will populate its cache from
cluster status as well as the key value store.