hostid not published or deleted causes events not reported to hamgr
bug https://platform9.atlassian.net/browse/IAAS-9344

this happened when the code tries to clean up already reported events
from kv store, but it uses a wrong key, which is randomly assigned
value when it check consul info. instead, it should be exactlly the
key for the report.

it was found by comparing the result of 'consul members' and the
result 'consul kv get --recurse' where there are active members are
not registered in kv store.